backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
google-generativeai>=0.8.3
python-dotenv==1.0.0
pytest==7.4.3
